+++ libs/components/ag-grid/src/lib/modules/ag-grid/ag-grid.service.ts
@@ -12,12 +12,12 @@ import {
   HeaderClassParams,
   ICellRendererParams,
   RowClassParams,
+  RowSelectionOptions,
   SuppressHeaderKeyboardEventParams,
   SuppressKeyboardEventParams,
   ValueFormatterParams,
 } from 'ag-grid-community';
-import { Subject } from 'rxjs';
-import { takeUntil } from 'rxjs/operators';
+import { Subject, takeUntil } from 'rxjs';
 
 import { SkyAgGridAdapterService } from './ag-grid-adapter.service';
 import { applySkyLookupPropertiesDefaults } from './apply-lookup-properties-defaults';
@@ -208,6 +208,26 @@ export class SkyAgGridService implements OnDestroy {
         providedGridOptions.cellSelection = true;
       }
       delete providedGridOptions.enableRangeSelection;
+    }
+    if (
+      providedGridOptions.rowSelection === 'single' ||
+      providedGridOptions.rowSelection === 'multiple'
+    ) {
+      const rowSelectionOptions: Record<string, RowSelectionOptions> = {
+        single: { mode: 'singleRow' },
+        multiple: { mode: 'multiRow' },
+      };
+      providedGridOptions.rowSelection =
+        rowSelectionOptions[providedGridOptions.rowSelection];
+    }
+    if (
+      defaultGridOptions.rowSelection &&
+      providedGridOptions.rowSelection?.mode
+    ) {
+      providedGridOptions.rowSelection = {
+        ...(defaultGridOptions.rowSelection as RowSelectionOptions),
+        ...providedGridOptions.rowSelection,
+      } as RowSelectionOptions;
     }
 
     const mergedGridOptions: GridOptions = {